Fix CI lint job: real ruff 0.16.3 default ruleset was far broader than local run showed
The previous commit's "lint" CI job failed (58 findings), even though ruff
had checked clean locally beforehand. Root cause: this sandbox had two ruff
installs -- an old 0.15.8 in ~/.local/bin shadowing the pinned 0.16.3 in
/usr/local/bin on PATH -- so every local `ruff check` silently ran the wrong
version with a much narrower default ruleset than what CI (correctly running
0.16.3) actually applied. python3 -m ruff (bypassing PATH) confirmed 0.16.3's
real defaults are broad -- effectively everything except a few opt-in-only
categories, not the classic minimal E4/E7/E9/F set older ruff versions used.

Rather than either accept that huge default wholesale or mechanically
silence findings, added a deliberate pyproject.toml [tool.ruff] config:
select = ["E", "F", "B", "C4", "I", "UP", "SIM", "RUF", "BLE", "ASYNC"],
line-length = 120 (matches this codebase's established style -- only a
handful of lines exceed it), and flake8-bugbear's extend-immutable-calls
set to recognize FastAPI's Depends()/Query() default-argument pattern as
the safe, idiomatic thing it is (not the mutable-default footgun B008
otherwise flags). DTZ (naive-datetime warnings) is deliberately excluded:
this app stores/compares naive UTC datetimes everywhere by design (see
app/util.py's utcnow(), app/schema.py's docstring), so requiring
tzinfo-aware datetimes would fight the codebase's own design, not fix a bug.

What was left after that (47 findings) was fixed for real, not suppressed:
- All 8 `class Foo(str, Enum)` models converted to Python 3.11's
  `enum.StrEnum` -- verified nothing relied on the old str+Enum's
  `str(member) == "ClassName.MEMBER"` quirk (StrEnum returns just the
  value), and re-verified via a live API smoke test that JSON responses
  still serialize as plain strings.
- A real async bug: app/adapters/onvif_ptz_bridge.py's watch() is `async
  def` but called blocking `time.sleep()` in its poll loop, which would
  stall the whole event loop the moment this bridge needs to run alongside
  anything else. Fixed to `await asyncio.sleep()`.
- ASYNC110 in astm_remote_id_ble_bridge.py: a `while True: await
  asyncio.sleep(1.0)` used only to keep a BLE scanner's context open
  indefinitely, replaced with the idiomatic `await asyncio.Event().wait()`.
- Two deliberately-broad `except Exception` catches (a UDP ingest loop
  parsing untrusted external datagrams; a DB liveness check) kept as-is
  but marked `# noqa: BLE001` with inline rationale, rather than either
  narrowed (risking silently missing real failure modes from a
  third-party parser) or the whole rule disabled project-wide.
- A SIM114 auto-fix had collapsed two clear if/elif branches in
  tracking.py's classification-upgrade logic into one 191-char unreadable
  line; reformatted into two named boolean conditions instead of reverting
  the merge outright.
- The rest: real unused-variable/dead-code findings (F841/RUF059), a
  redundant int(round(...)) double-cast verified empirically safe with
  numpy scalars, unnecessary dict()-call-as-literal rewrites, an unsorted
  import block, and line wraps for the genuinely-long remaining lines.

Verified: ruff and mypy both clean; full test suite (335 tests) passing on
both SQLite and PostgreSQL; a live API smoke test confirming StrEnum JSON
serialization is unaffected.

+++ app/acoustic_beamforming.py
@@ -90,7 +90,7 @@ def _steered_power(
     summed = np.zeros(channels.shape[1])
     for i in range(n_mics):
         delay_s = steering_delay_s(mic_positions_m[i], azimuth_deg)
-        shift_samples = int(round(delay_s * sample_rate_hz))
+        shift_samples = round(delay_s * sample_rate_hz)
         # A source at `azimuth_deg` reaches this mic `delay_s` earlier/later
         # than the array center; rolling the *opposite* direction realigns
         # it back to the center's timeline before summing.
